+++ demisto_sdk/commands/common/hook_validations/description.py
@@ -9,6 +9,7 @@ from demisto_sdk.commands.common.hook_validations.base_validator import (
     BaseValidator, error_codes)
 from demisto_sdk.commands.common.hook_validations.structure import \
     StructureValidator
+from demisto_sdk.commands.common.mardown_lint import has_markdown_lint_errors
 from demisto_sdk.commands.common.tools import find_type, get_yaml, os, re
 
 CONTRIBUTOR_DETAILED_DESC = 'Contributed Integration'
@@ -38,6 +39,7 @@ class DescriptionValidator(BaseValidator):
 
         # make sure the description is a seperate file
         if not self.data_dictionary.get('detaileddescription') and '.md' in self.file_path:
+            self.has_markdown_lint_errors()
             self.is_valid_description_name()
             self.contains_contrib_details()
 
@@ -219,4 +221,13 @@ class DescriptionValidator(BaseValidator):
                 self._is_valid = False
                 return False
 
+        return True
+
+    @error_codes('DS108')
+    def has_markdown_lint_errors(self):
+        if has_markdown_lint_errors(self.file_path):
+            error_message, error_code = Errors.description_lint_errors(self.file_path)
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                self._is_valid = False
+                return False
         return True