+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed an issue where the **doc-review** command showed irrelevant messages.
 * Fixed an issue in **validate**, where backward-compatibility failures prevented other validations from running.
 * fixed an issue in **validate**, where content-like files under infrastructure paths were not ignored.
+* Added support for linting markdown files in the **validate** command.
 
 ## 1.7.3